chore: update from code review
Resolves (#34)
